pyproject.toml
[tool.poetry]
name = "autonomous-ai-challenge"
version = "0.1.0"
description = ""
authors = ["unerue <unerue@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
torch = {version = "2.3.1+cu121", source = "pytorch"}
torchvision = {version = "0.18.1+cu121", source = "pytorch"}
ultralytics = "^8.3.8"
albumentations = "^1.4.18"
transformers = "^4.45.2"
tqdm = "^4.66.5"
numpy = "^2.1.2"
pillow = "^10.4.0"
pycocotools = "^2.0.8"
timm = "^1.0.11"
torchmetrics = "^1.6.0"
mlflow = "^2.18.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
